remove 32 bit datatypes

+++ src/bloom.c
@@ -18,13 +18,13 @@
  * Returns:
  *     unsigned integer
  */
-static uint32_t ideal_size(const uint32_t expected, const float accuracy) {
+static uint64_t ideal_size(const uint64_t expected, const float accuracy) {
 	return -(expected * log(accuracy) / pow(log(2.0), 2));
 }
 
 // TODO: specify which hashing algorithm to use.
 // TODO: add comment/documentation
-bool bloom_init(bloomfilter *bf, const uint32_t expected, const float accuracy) {
+bool bloom_init(bloomfilter *bf, const uint64_t expected, const float accuracy) {
 	bf->size        = ideal_size(expected, accuracy);
 	bf->hashcount   = (bf->size / expected) * log(2);
 	bf->bitmap_size = ceil(bf->size / 8);
